refactor(osio): Move OsioStatusEventKind to base package
