+++ addons/osio-addon/src/main/java/io/fabric8/launcher/osio/OsioStatusEventKind.java
@@ -1,4 +1,4 @@
-package io.fabric8.launcher.osio.steps;
+package io.fabric8.launcher.osio;
 
 import io.fabric8.launcher.core.api.events.StatusEventKind;
